+++ Controller/ChangePasswordController.php
@@ -16,6 +16,67 @@ use FOS\UserBundle\Controller\ChangePasswordController as BaseController;
 
 class ChangePasswordController extends BaseController
 {
+//This method has to be removed if https://github.com/FriendsOfSymfony/FOSUserBundle/pull/2587 is merged
+    /**
+     * Change user password.
+     *
+     * @param Request $request
+     *
+     * @return Response
+     */
+    public function changePasswordAction(Request $request)
+    {
+        $user = $this->getUser();
+        if (!is_object($user) || !$user instanceof UserInterface) {
+            throw new AccessDeniedException('This user does not have access to this section.');
+        }
+
+        /** @var $dispatcher EventDispatcherInterface */
+        $dispatcher = $this->get('event_dispatcher');
+
+        $event = new GetResponseUserEvent($user, $request);
+        $dispatcher->dispatch(FOSUserEvents::CHANGE_PASSWORD_INITIALIZE, $event);
+
+        if (null !== $event->getResponse()) {
+            return $event->getResponse();
+        }
+
+        /** @var $formFactory FactoryInterface */
+        $formFactory = $this->get('fos_user.change_password.form.factory');
+
+        $form = $formFactory->createForm();
+        $form->setData($user);
+
+        $form->handleRequest($request);
+
+        if ($form->isSubmitted() && $form->isValid()) {
+            /** @var $userManager UserManagerInterface */
+            $userManager = $this->get('fos_user.user_manager');
+
+            $event = new FormEvent($form, $request);
+            $dispatcher->dispatch(FOSUserEvents::CHANGE_PASSWORD_SUCCESS, $event);
+
+            $userManager->updateUser($user);
+
+            if (null === $response = $event->getResponse()) {
+                $url = $this->generateUrl('fos_user_profile_show');
+                $response = new RedirectResponse($url);
+            }
+
+            $dispatcher->dispatch(FOSUserEvents::CHANGE_PASSWORD_COMPLETED, new FilterUserResponseEvent($user, $request, $response));
+
+            return $response;
+        }
+
+        return $this->render('@FOSUser/ChangePassword/change_password.html.twig', array(
+            'form' => $form->createView(),
+            'data' => $this->setUserData(),
+        ));
+    }
+
+    /*
+     * Method to override to send data to the template
+     */
     public function setUserData()
     {
         return array(