+++ Controller/ProfileController.php
@@ -16,7 +16,83 @@ use FOS\UserBundle\Controller\ProfileController as BaseController;
 
 class ProfileController extends BaseController
 {
-    public function setUserData()
+//This method has to be removed if https://github.com/FriendsOfSymfony/FOSUserBundle/pull/2587 is merged
+    /**
+     * Show the user.
+     */
+    public function showAction()
+    {
+        $user = $this->getUser();
+        if (!is_object($user) || !$user instanceof UserInterface) {
+            throw new AccessDeniedException('This user does not have access to this section.');
+        }
+
+        return $this->render('@FOSUser/Profile/show.html.twig', array(
+            'user' => $user,
+            'data' => $this->setUserData(),
+        ));
+    }
+//This method has to be removed if https://github.com/FriendsOfSymfony/FOSUserBundle/pull/2587 is merged
+    /**
+     * Edit the user.
+     *
+     * @param Request $request
+     *
+     * @return Response
+     */
+    public function editAction(Request $request)
+    {
+        $user = $this->getUser();
+        if (!is_object($user) || !$user instanceof UserInterface) {
+            throw new AccessDeniedException('This user does not have access to this section.');
+        }
+
+        /** @var $dispatcher EventDispatcherInterface */
+        $dispatcher = $this->get('event_dispatcher');
+
+        $event = new GetResponseUserEvent($user, $request);
+        $dispatcher->dispatch(FOSUserEvents::PROFILE_EDIT_INITIALIZE, $event);
+
+        if (null !== $event->getResponse()) {
+            return $event->getResponse();
+        }
+
+        /** @var $formFactory FactoryInterface */
+        $formFactory = $this->get('fos_user.profile.form.factory');
+
+        $form = $formFactory->createForm();
+        $form->setData($user);
+
+        $form->handleRequest($request);
+
+        if ($form->isSubmitted() && $form->isValid()) {
+            /** @var $userManager UserManagerInterface */
+            $userManager = $this->get('fos_user.user_manager');
+
+            $event = new FormEvent($form, $request);
+            $dispatcher->dispatch(FOSUserEvents::PROFILE_EDIT_SUCCESS, $event);
+
+            $userManager->updateUser($user);
+
+            if (null === $response = $event->getResponse()) {
+                $url = $this->generateUrl('fos_user_profile_show');
+                $response = new RedirectResponse($url);
+            }
+
+            $dispatcher->dispatch(FOSUserEvents::PROFILE_EDIT_COMPLETED, new FilterUserResponseEvent($user, $request, $response));
+
+            return $response;
+        }
+
+        return $this->render('@FOSUser/Profile/edit.html.twig', array(
+            'form' => $form->createView(),
+            'data' => $this->setUserData(),
+        ));
+    }
+
+    /*
+     * Method to override to send data to the template
+     */    public function setUserData()
     {
         return array(
             'gravatar' => $this->getParameter('c975_l_user_files.gravatar'),